+++ pybamm/__init__.py
@@ -125,7 +125,7 @@ from .models.event import Event
 from .models.event import EventType
 
 # Battery models
-from .models.full_battery_models.base_battery_model import BaseBatteryModel
+from .models.full_battery_models.base_battery_model import BaseBatteryModel, Options
 from .models.full_battery_models import lead_acid
 from .models.full_battery_models import lithium_ion
 
@@ -252,10 +252,16 @@ default_plot_style = os.path.join(root_dir(), "pybamm/plotting/pybamm.mplstyle")
 import matplotlib.pyplot as plt
 
 plt.style.use(default_plot_style)
+
 #
 # Simulation
 #
 from .simulation import Simulation, load_sim, is_notebook
+
+#
+# Batch Study
+#
+from .batch_study import BatchStudy
 
 #
 # Remove any imported modules, so we don't expose them as part of pybamm